Use multi match queries
Start by weighting readmes twice as high as descriptions. Descriptions are
really short, and it seems like this causes elasticsearch to think it's a good
match when a couple of words in the query match (there is no minimum_should_match yet)

I'm not bothering to include to organisations because those fields only convey
authorship and I want to handle that with facets.