Separate app deployment from rest of the playbook
The deployment part is really hacky as it just rsyncs code. This could be done
a lot better, and I only want to run it when pushing stuff to the cloud.

On the other hand the rest of it provisions using packaged software. It should
be safe to run any time.

Added a Procfile so it's easier to run the rails app on the VM, and upgraded to
a newer ubuntu so dev is more similar to live.